refactor: cleanup types
