+++ lumigator/frontend/src/helpers/retrieveEntrypoint.ts
@@ -1,4 +1,4 @@
-import type { JobDetails } from '@/types/JobDetails'
+import type { Job } from '@/types/Job'
 
 /**
  * Retrieves and normalizes the entrypoint configuration from a job object.
@@ -7,7 +7,7 @@ import type { JobDetails } from '@/types/JobDetails'
  * @param {string} job.entrypoint - The entrypoint stringified JSON provided from the API
  * @returns {Object|null} The parsed and normalized entrypoint configuration object, or null if invalid.
  */
-export function retrieveEntrypoint(job: JobDetails) {
+export function retrieveEntrypoint(job: Job) {
   if (!job || !job.entrypoint) {
     console.error('Invalid job data')
     return
