+++ lumigator/frontend/src/sdk/workflowsService.ts
@@ -1,5 +1,6 @@
 import { lumigatorApiAxiosInstance } from '@/helpers/lumigatorAxiosInstance'
-import type { CreateWorkflowPayload, Workflow, WorkflowResults } from '@/types/Workflow'
+import type { WorkflowResults } from '@/types/Metrics'
+import type { CreateWorkflowPayload, Workflow } from '@/types/Workflow'
 
 async function createWorkflow(experimentPayload: CreateWorkflowPayload): Promise<Workflow> {
   const response = await lumigatorApiAxiosInstance.post('workflows', experimentPayload, {
