+++ lumigator/frontend/src/stores/experimentsStore.ts
@@ -9,14 +9,15 @@ import type { EvaluationJobResults, ExperimentResults } from '@/types/Experiment
 import { downloadContent } from '@/helpers/downloadContent'
 import type { Model } from '@/types/Model'
 import { workflowsService } from '@/sdk/workflowsService'
-import type { ExperimentNew } from '@/types/ExperimentNew'
-import { WorkflowStatus, type Workflow, type WorkflowResults } from '@/types/Workflow'
+import type { Experiment } from '@/types/Experiment'
+import { WorkflowStatus, type Workflow } from '@/types/Workflow'
 import axios from 'axios'
+import type { WorkflowResults } from '@/types/Metrics'
 
 export const useExperimentStore = defineStore('experiments', () => {
-  const experiments: Ref<ExperimentNew[]> = ref([])
+  const experiments: Ref<Experiment[]> = ref([])
 
-  const selectedExperiment: Ref<ExperimentNew | undefined> = ref()
+  const selectedExperiment: Ref<Experiment | undefined> = ref()
   const selectedWorkflow: Ref<Workflow | undefined> = ref()
   const selectedWorkflowResults: Ref<EvaluationJobResults[] | undefined> = ref()
 
@@ -37,7 +38,7 @@ export const useExperimentStore = defineStore('experiments', () => {
   }
 
   // aggregates the experiment's status based on its workflows statuses
-  function retrieveStatus(experiment: ExperimentNew): WorkflowStatus {
+  function retrieveStatus(experiment: Experiment): WorkflowStatus {
     const workflowStatuses = experiment.workflows.map((workflow) => workflow.status)
     const uniqueStatuses = new Set(workflowStatuses)
 
@@ -58,7 +59,7 @@ export const useExperimentStore = defineStore('experiments', () => {
    * The retrieved IDs will determine which experiment is still Running
    * @returns {string[]} IDs of stored experiments that have not completed
    */
-  function getIncompleteExperiments(): ExperimentNew[] {
+  function getIncompleteExperiments(): Experiment[] {
     return experiments.value.filter((experiment) => !completedStatus.includes(experiment.status))
   }
 
@@ -66,7 +67,7 @@ export const useExperimentStore = defineStore('experiments', () => {
    *
    * @param {string} id - String (UUID) representing the experiment which should be updated with the latest status
    */
-  async function updateExperimentStatus(experiment: ExperimentNew): Promise<void> {
+  async function updateExperimentStatus(experiment: Experiment): Promise<void> {
     try {
       const incompleteWorkflows = experiment.workflows.filter(
         (workflow) => !completedStatus.includes(workflow.status),
@@ -143,7 +144,7 @@ export const useExperimentStore = defineStore('experiments', () => {
     downloadContent(blob, `${selectedWorkflow.value?.name}_results`)
   }
 
-  async function fetchExperimentResults(experiment: ExperimentNew) {
+  async function fetchExperimentResults(experiment: Experiment) {
     for (const workflow of experiment.workflows) {
       //   console.log({experiment, workflow})
       if (workflow.artifacts_download_url) {
