+++ lumigator/frontend/src/types/Workflow.ts
@@ -1,5 +1,3 @@
-import type { Bertscore, Meteor, Rouge } from './Experiment'
-
 export type Workflow = {
   id: string
   experiment_id: string
@@ -32,63 +30,6 @@ export type Metrics = {
   rougeL_mean: number
   rougeLsum_mean: number
   rouge2_mean: number
-}
-
-export type WorkflowResults = {
-  artifacts: Artifacts
-  metrics: MetricsResult
-  parameters: Parameters
-}
-
-export type Artifacts = {
-  examples: string[]
-  ground_truth: string[]
-  model: string
-  predictions: string[]
-
-  summarization_time: number
-  evaluation_time: number
-}
-
-export type MetricsResult = {
-  bertscore: Bertscore
-  meteor: Meteor
-  rouge: Rouge
-}
-
-export type Parameters = {
-  dataset: {
-    path: string
-  }
-  evaluation: {
-    max_samples: number
-  }
-  metrics: string[]
-  return_input_data: boolean
-  return_predictions: boolean
-  storage_path: string
-  hf_pipeline: HfPipeline
-  inference_server?: unknown
-  name: string
-  job: {
-    enable_tqdm: boolean
-    max_samples: number
-    output_field: string
-    storage_path: string
-  }
-  params: unknown
-}
-
-export type HfPipeline = {
-  accelerator: string
-  max_new_tokens: number
-  model_uri: string
-  revision: string
-  task: string
-  torch_dtype: string
-  truncation: boolean
-  trust_remote_code: boolean
-  use_fast: boolean
 }
 
 export enum WorkflowStatus {
